LayoutTests/imported/w3c/web-platform-tests/dom/nodes/Document-constructor-expected.txt


FAIL new Document(): interfaces assert_false: Should not be an XMLDocument expected false got true
PASS new Document(): children 
PASS new Document(): metadata 
PASS new Document(): characterSet aliases 
FAIL new Document(): URL parsing assert_equals: expected "http://example.org/?%C3%A4" but got "http://example.org/?ä"
